Subs 1.0.0.0 -> 198.18.1.[1|2] in network checker
It turned out that some network equipment filters
out those fake addresses 1.0.0.0 which we used for
network checker probing packages. 198.18.1.0/24 is
one of those ranges which are reserved for benchmarking
network connectivities http://tools.ietf.org/html/rfc2544,
so it seems to be appropriate to use
those addresses for this case by default.

Change-Id: I616b524db0675d752279bdf294aaba486d253336
Closes-Bug: #1291655
Signed-off-by: Dmitry Borodaenko <dborodaenko@mirantis.com>