Confirm disk re-partitioning before proceeding
Ask for a confirmation before Fuel installation on a drive with
existing partition table.

New kickstart kernel parameter:
  - "forceformat" - (yes/no) defaul is "no". Set to "yes" if you
                    want to format chosen drive automatically.

For example, if you add "installdrive=sda forceformat=yes" to the
kernel parameters line in Grub menu, then Fuel "sda" drive will be
formatted automatically and Fuel will be installed on it.

Change-Id: I0dfcb1d461748fe6b0698f125201b730d63e41d4
Closes-bug: #1325068