Add yum.rb provider for 5.0 and 5.0.1 with fixes
We found a critical bug in puppet's versioncmp function which deosn't
behave like RPM does, so we've implemented a new function as a fix.

Now we have to deliver this fix for both 5.0 and 5.0.1, because without
this fix the rollback to 5.0 and 5.0.1 will fails.

Related patch for fuel_upgrade script:

    https://review.openstack.org/#/c/111238/

Change-Id: Ic456a6f26fb2c33f4e8cddb3737ab3e0bca690e3
Blueprint: upgrade-to-5-1