Missed renaming for Sahara system tests
We can still see the word "savanna" in the names of variables, methods
and classes in system tests for Sahara. This fact causes errors when
running OSTF tests because all variables, methods and classes have been
renamed in OSTF tests for Sahara.

In addition, variable SERVTEST_SAVANNA_SERVER_URL has been removed 
because there is no code fragment where this variable is used.

Closes-Bug: #1353433

Change-Id: Ia414f9f5bbb0a5efbab537f7f2a1b381dc7387ca