Add fuel-centos-build container for fuel/centos
fuel/centos will now be built from CentOS6
container dynamically with current package set
instead of being periodically rebuilt. It uses
ami-creator, which only runs on CentOS. It is
run inside Docker so it can be used on an
Ubuntu build host. After the image is created,
img2docker loads the content into Docker as a
Docker image.

Change-Id: Ibff08d309fe091cc40d053781b7556fd36ebf173
Closes-Bug: #1336911