Add method which returns cluster public VIP by id
When we want to connect to OpenStack services API which
are running on HA cloud, we need to have ability to get public
virtual IP address of deployed environment. So adding an
additional method to the FuelWebClient class for that.

Related-bug: #1369412
Change-Id: I98c8c323dd81f0b361b73050df29083e0d912211
(cherry picked from commit 146234cfa742fa73dcd3d41798297ff16497027e)