requirements-deb: skip the obosolete software
Skip the dependencies of the very old qemu (pre 1.0) versions. Same for
the pacemaker <= 1.1.6. The download script (mkrepo.sh) picks the
correct libraries on its own (in fact it has been downloading for a long
time so none bothered to update the dependencies).

While at it remove Linux 3.2 packages too. These haven't been actually
used for a long time (starting from Fuel 4.x?), however it has been
installed due to a bug in cobbler configuration. That bug has been fixed
now (see https://review.openstack.org/#/c/135677), so it's time to purge
the unused Linux 3.2 packages.

Last but not least this patch helps to update Ubuntu to 14.04 (the packages
excluded by this patch are not available in the newer versions of Ubuntu).

Related-bug: #1383641
blueprint support-ubuntu-trusty
Change-Id: I4f278d08ff643c7bb90d431c7cee4f79a72d32a1