Remove running of one Heat test in system test 'Heat HA'
Heat OSTF 'test_autoscaling' is skipped in the HA mode, but in system
tests skipped test equal failed test, it is not correct in current situation. So need to don't run test 'test_autoscaling' in the HA Heat system test.

This fix was already done in 5.1 in the following commit:
https://review.openstack.org/#/c/122364/ , but it was accidentally
reverted, so need to add it again.

Change-Id: Ia079ead1ad35c5a65a8139f0dbcd1b4731101dd9
Closes-Bug: #1371025