ubuntu: simplify the kernel updates
At the moment the kernel version to be installed on Ubuntu nodes is hard
coded in config.mk (and several other files, such as requirements-deb.txt).
This version must coninside with the version of the kernel used by Debian
installer (otherwise Debian installer won't be able to load any kernel
modules which in general breaks the provisioning). As a result updating
the kernel (say, to incorporate the upstream security fixes) requires
a lot of manual intervention: downloading the netboot image, finding out
the exact kernel version it uses, updating config.mk, requirements-deb.txt,
and so on.

This patch makes the process much simpler: updating the kernel (to a version
provided by Ubuntu) boils down to changing the UBUNTU_KERNEL_FLAVOR
variable (and possilby the UBUNTU_NETBOOT_FLAVOR variable).

 * config.mk: introduce UBUNTU_KERNEL_FLAVOR and UBUNTU_NETBOOT_FLAVOR
   variables. The former specifies which kernel should be installed on
   nodes, and the latter tells which netboot variant should be used.

 * image/ubuntu/module.mk: install the kernel specified by
   the UBUNTU_KERNEL_FLAVOR variable. There's no need to use exactly
   the same version of the kernel as Debian installer does, as image
   based provisioning does not use Debian installer at all.

 * mirror/ubuntu/boot.mk: guess and write the installer kernel version
   into a file which can be include'd by other modules. While at it
   reduce the copy-paste noise.

 * mkrepo.sh: make sure that the kernel (and the correspodning headers)
   (meta-)packages specified by UBUNTU_KERNEL_FLAVOR are included into
   the mirror.

Change-Id: Ia8db615284644b0e2bcdb70c1c7bb1f4388d5e46
Related-bug: #1390044
Related-bug: #1383641