mirror-ubuntu: use the OSCI repo for bootstrapping the chroot
If some of the base packages (like python-minimal) has been rebuilt
we should use them when bootstrapping the temporary chroot. Otherwise
the newly created apt repo will contain either only upstream version of
the package in question, or both the upstream and customized versions.

Assume that the base package X has been customized. The customized version
of X is located on the OSCI mirror. Multistrap does not use that mirror, so
the upstream version of X gets pulled into the apt cache (typically
/var/apt/cache/archives) during bootstrapping of the chroot. Packages
depending on X pull its customized version (since the OSCI repo is added
to the sources.list.d after bootstrapping the chroot). If X is an essential
package (so other packages are not obliged to declare dependence on X)
only the upstream version of X gets into the apt repo being created.

Use the OSCI mirror when bootstrapping the chroot to avoid the problem.

While at it reduce copy-paste noise in createchroot.mk and make it a little
bit less obscure.

Related-bug: #1390416
Related-bug: #1390044
Related-bug: #1383641

Change-Id: Ic418c760a1635d5c72cc108658d412228fb83a62