packages/deb: build packages using the correct mirror (LP #1390416)
As of now Fuel related deb packages are built using the upstream mirror.
This yeilds inconsistent dependencies, for instance, packages depend on
both the upstream and the customized python versions, and provisioning of
Ubuntu node fails as a result.

To resolve the problem setup the build chroot using the official Fuel apt
repo. While at it use the locally created apt mirror instead of downloading
packages once more, and fix a number of inconsistencies/bugs.

* sandbox.mk:
  - sandbox define is modified so as to use local pre-built mirror.

* packages/deb/module.mk:
  - use the helper script to create build sandbox.

* requirements-deb.txt:
  - make sure tools required for building Debian packages are included
    into the local apt repo.

Change-Id: I9f7b06bba94a65b2d65168cc1122257f4bec78be
Related-bug: #1390416