ubuntu: produce a valid Debian installer initrd.gz
The code which applies initrd patch(es) is produces a half-broken initrd
which is a) not compressed, b) contains a chunk of itself:

$ initrd=/mnt/ubuntu/installer-amd64/current/images/netboot/ubuntu-installer/amd64/initrd.gz
$ file $initrd
/mnt/ubuntu/installer-amd64/current/images/netboot/ubuntu-installer/amd64/initrd.gz: ASCII cpio archive (SVR4 with no CRC)
$ cpio --list < $initrd | grep -e 'initrd\.gz'
initrd.gz

Fortunately this initrd (or rather initramfs) is still valid, however, it's
2.5x larger than the original. This slows down provisioning for no good
reason and puts an extra load on the master node.

Make sure the modified initrd is compressed and does not contain itself.

Closes-bug: #1397845
Change-Id: I666a4395e577b917016c2220a760a4cf340a54d2