ubuntu: produce a valid Debian installer initrd.gz
The code which applies initrd patch(es) produces a half-broken initrd which
is a) not compressed, b) contains a chunk of itself. Fortunately the initrd
is still able to boot, however, it's 2.5x larger than the original. This
slows down the provisioning and puts an extra load on the master node.

Closes-bug: #1397845
Change-Id: I666a4395e577b917016c2220a760a4cf340a54d2