Add check for DHCP lease that remains after deletion the instance
1) With nova-network, when an instance is migrated from one compute
node to another, it's DHCP lease remains on the first compute even
after instance is deleted.
This makes it impossible to obtain the same address for a new
instance with another MAC address which is created on the first
compute node.

The check will fail the test if the DHCP lease still remains on the
first compute node after instance migration.

2) Added @retry decorator to the method execute_through_host()

3) Fixed issue with ping 8.8.8.8 (address wasn't provided to the
command)

Change-Id: I77c0c4b540af880f81076a59109f7affc0e5f628
Related-Bug: #1391010