Fixing some issue in logging
The method "assert_ostf_run_certain" in fuel_web_client.py
takes a list of tests that must have succes as one of parameters.
If the list has more than 1 item, method "assert_ostf_run_certain"
will not work correctly. For example, if we have three tests in the 
list, only the last item of this list will always be passed into 
function for logging.

Change-Id: Id2c03a874aa59ba15cd6697228c895d5ecff47e1