mirror/ubuntu: make the deb package download easier to debug
At the moment debugging package download issues (finding out why certain
package is missing or has the dependency problems, etc) involves a lot of
guesswork sinse the temporary chroot gets removed. Apparently the only for
removing is that the directory holding the chroot gets copied to the ISO.
Create the chroot in the `build' directory rather than in local_mirror so
that removing the chroot is not mandatory any more.

* createchroot.mk:
 - introduce CHROOT_DIR (per target) variable and make chroot there,
   put CHROOT_DIR into BUILD_DIR.
 - copy only useful files (pool, dists, indices) to the local_mirror
 - don't remove the chroot, it might be very useful for debugging.

* mkrepo.sh: store APT configs/state in a different directory in order
  to not pollute the ISO with temporary files

blueprint support-ubuntu-trusty

Change-Id: Ia10f82dc79933e41a489b6e4f5d5523ce3c50d9a