The vcenter_vmdk system test expects to fail two OSTF tests
The system test 'vcenter_vmdk' expected to fail an only OSTF test
'Launch instance, create snapshot, launch instance from snapshot'.
But because of wrong CirrOS image an another OSTF fails too:
'Create volume and attach it to instance'. Both of them should be expected
as failed in that test case.

Change-Id: I5ec0bf366409a69d0a1cfd2ef041318a6fee6a10
Closes-bug: #1417679