Change nodes so that they end in admin mac suffx instead of number
Previously the nodes would end in a number iterator to ensure they where
unique, however this makes working with the nodes generated this way
difficult as they are referred to as by their admin interface mac.

This changes it so the slave nodes will receive the suffix of the admin
mac.

Change-Id: I74743156c92cfeb9ec7a54f6cced65fd927ec2e7