Added fuel-provisioning-scripts package
This package is going to provide a set of executable
scripts which we need for provisioning to make it work.

At this iteration we only put download-debian-installer
script which is going to be used for downloading ubuntu debian-installer
kernel and initrd from user defined repositories.

Depends on: Ic6ffc996fe019628df12da359e76cc17bb963b13
Change-Id: I6628d85c681f21269cba157fc2650e45d730fbb9
Implements: blueprint consume-external-ubuntu