Added fuel-provisioning-scripts package
This package is going to provide a set of executable
scripts which we need for provisioning to make it work.

At this iteration we only put download-debian-installer
script which is going to be used for downloading ubuntu debian-installer
kernel and initrd from user defined repositories.

Depends on: I9e7ac50d19ea8b984e1ad2831c73f2cb88ddbd54
Change-Id: I6628d85c681f21269cba157fc2650e45d730fbb9
Implements: blueprint consume-external-ubuntu