Removed debian-installer initrd and kernel
We are not allowed to re-distribute ubuntu debian-installer
kernel and initrd. So, we are going to download these files
from one of user defined ubuntu repos. We implemented this
as a pre-provision task which is supposed to run a shell
script.

This patch removes original ubuntu debian-installer
ramdisk and kernel and substitutes them with zero files just
to make sure cobbler is able to create corresponding Distro.

Change-Id: Ie154cfdf3c6e9b4f063ab859cd050be6faa1707e
Implements: blueprint consume-external-ubuntu