requirements-deb: re-add pacemaker configuration system (pcs)
The python-pcs package is the packet construction set [1]. This package
has been shipped with Debian/Ubuntu for quite a long time. For some
reason the pacemaker/corosync configuration system [2] has been packaged
for Ubuntu under the name `python-pcs'. Use the correct name (that is,
`pcs', like its RPM counterpart) to resolve the collision.

[1] https://github.com/gvnn3/PCS
[2] https://github.com/feist/pcs

Note: this patch should be merged along with the following ones:
https://review.fuel-infra.org/4522
https://review.openstack.org/164109

blueprint support-ubuntu-trusty

Change-Id: I7291aff00de3a69add5a2ce1b4edff3bd33c1ba8