IBP: fix apt pinning in build script
We have to use `archive` name in APT pinning because `suite` is actually
an archive name. For instance, the following upstream Ubuntu repo

    deb http://archive.ubuntu.com/ubuntu/ trusty-updates main

has the following attributes in APT pinning

    a=trusty-updates, n=trusty, c=main

Obviously, it's wrong to use `n=` (codename) for matching suite.

Besides this, since now we're generating a few pinnings for the source
like

    deb http://archive.ubuntu.com/ubuntu/ trusty main restricted

because the following line

    a=trusty, c=main, c=restricted

will apply pinning only for the `restricted` section.

Also, it skips a repo from pinning if its priority isn't set.

Change-Id: Ibe5c99910f70312ea4e4bd0c642cd8f31ddc526b
Implements: blueprint consume-external-ubuntu
Closes-Bug: #1436806