Added priorities for centos repos
We need to be able to download packages from
various repositories taking into account repository
priority. We need this for patching story.

This patch introduces MULTI_MIRROR_CENTOS variable
which is supposed to have the following format:

  "name1,url1,pri1 name2,url2,pri2"

In case when pri is not defined it will be set to 10,
for example
  "name1,url1"
        is equal to:
  "name1,url1,10"

Besides, this patch introduces uniformity for all those
cases USE_MIRROR=msk|srt|none removing unused variables.

It is supposed that USE_MIRROR=none is only used for
building Centos mirror not for building ISO.

Change-Id: I84a893a32194f0d6f5c2727af373f680fae1aaae
Closes-Bug: #1446293