mkrepo.sh: download base packages properly.
Skip the epoch in the packages' file names in order to not confuse
debootstrap (which fails if the deb package file name contains the colon
character). In order to avoid the problem download the base packages
in the same manners as others instead of copying them from multistrap
cache (it's multistrap which modifies the packages' file names).

Partial-Bug: #1441611
Change-Id: Ib5b857a0ae637cc5c327e45f28e8397ae000b78f