Add updates and security repositories
Added yum repositories for mos?.?-updates
and mos?.?-security, depending on Fuel version

After running fuelmenu, updates repo will be
verified that it can be resolved and download
updates. If updates fail or resolution fails,
a warning is displayed at the end of deployment
and broadcast to Fuel UI.

This point was selected because Internet access
is usually not available until after fuelmenu
has been run.

Change-Id: I30e8c854147bed9901470bf1a98875180d808ab4
Partial-Bug: #1466080
Partial-Bug: #1466070