Added priorities for centos extra repos
We need to be able to download packages from
various repositories taking into account repository
priority. We need this for patching story.

By default priorities are
1) upstream 90
2) proprietary 20

EXTRA_RPM_REPOS is to have the following format
"name1,url1,pri1 name2,url2,pri2"

In case when pri is not defined it will be set to 10,
for example "name1,url1" is equal to "name1,url1,10"

Closes-Bug: #1446293
Change-Id: I4a26dd005852af60104c39e46384a21c39cce587