Wrong requirement on fuel-library in package fuel
Fuel library packages have names in format 'fuel-libraryN.N'. This
is required for having possibility to install different versions of
'fuel-library' packages to masternode. In requirements for package
'fuel' we have requrement on 'fuel-library >= 7.0', but version is
bound to name 'fuel-libraryN.N'. Both 'fuel-library' packages
for 6.1 and 7.1 provide 'fuel-library'. So then package 'fuel'
checks requrement on 'fuel-library >= 7.0' it satisfied by any
installed 'fuel-library' package. In case of upgrade from 6.1 to
7.0 we don't have fuel-library7.0 installed.

Change-Id: Ic0056b187a95e1afa45edd0b67e0453d7c557edd
Closes-Bug: #1474883