bootstrap_admin_node: improve check for Ubuntu mirror
The master node deployment script (bootstrap_admin_node.sh) starts early
in the boot sequence. At this moment the network configuration might be
still in progress (especially if some interfaces are configured via DHCP).
Therefore the check for Ubuntu mirror accessibility can fail and pop up
Fuel menu which breaks automatic deployment. In order to avoid the problem
repeat the check several times, and show fuel menu if the mirror is
still not available in a minute.

Closes-Bug: #1477980
Change-Id: I65077e159f6ab4592b29342fa08363f9279dc8a2