Add sleep for ext config, refactor bootstrap_admin config
This unbinds the waiting period for enabling external configuration
of Fuel Master and fuelmenu. Now after fuelmenu runs (enabled or
not), there is an optional 3000s sleep that is activated when Fuel
is installed with wait_for_external_config=yes as a kernel parameter.

Also refactored showmenu option to use a new config file
/etc/fuel/bootstrap_admin_node.conf.

Change-Id: I08686ced0b43ed35c5fcbbe4d2c9102b3483ee0b
Partial-Bug: #1461052