Fix file system mount in the future
The issue is that when image is built and immediately
copied to a slave node and if the slave node is not
synchronized with master node, we can encounter
the error when running e2fsck

Superblock last write time is in the future.

And then e2fsck returns exit code 1 which leads
fuel agent to fail.

broken_system_clock = true e2fsck option fixes the issue.

Change-Id: If8db3c1b09521d4093fdc928d595240928e92e49
Closes-Bug: #1489475
Co-Authored-By: Alexey Sheplyakov <asheplyakov@mirantis.com>