bootstrap: force dhclient to restart configuration on failures
dhclient shipped with CentOS bails out (after sending DHCPDECLINE) if
a duplicate address has been detected and `-1' ("one shot") flag has
been given. This kind of violates the RFC 2131 which prescribes
the client to restart the configuration process. On the other hand
retrying the configuration denies the `-1' switch purpose. CentOS'
developers choose to violate the RFC to obey the "one shot" behavior.
As a result network-scripts are unable to handle IP addresses conflict
and fail to bring the interfaces up during the OS boot sequence.

Suppose there are some bootstrap nodes booted by a different master
node (say, by a previously deployed one) in the admin/pxe network.
The DHCP server on the current (the newly deployed) master node is
not aware of those nodes, thus the IP addresses leased to the nodes
collide with the ones of the (stale) nodes booted by a different
master node. As a result some nodes will fail to boot.

In order to solve the problem specify the PERSISTENT_DHCLIENT flag
in the ifcfg-${boot_iface} config file, so dhclient will retry
the configuration.

Closes-Bug: #1492307
Change-Id: Ica89fc05d4f936feb3066d27a8fb9a24bc36b55c