Better message about updates available
Message used can easily scare off: if Fuel installs without
errors, there are should not be any warnings. We had warning
even if both installation is successful and Internet is
available for updates. I propose to change it to notice
(same type of message which is used for "Master node installation has
been completed successfully").

Change-Id: If7911f9a28d7d1d5977819a01612b051cd4e0e72
Related-Bug: #1495658