Swap manifest location check order in mcollective container
Currently Dockerfile checks if "old" manifest exists, and if not,
proceeds to the to-be-introduced one. This behavior makes it
impossible to make CI pass for I1657ce22be62414abd4a407a5d28292f25915a7b.

The modified conditional will fallback to the old manifest only if new
one doesn't exist.

Related-bug: 1508058
Change-Id: Iab069b6c660edf0c7932a61cc480530447932708