Update dependency checks
VirtualBox scripts rely on various utilities to function. To validate
the environment, virtualbox/prepare-environment.sh checks if all these
utilities are available to the shell. However, some of the checks
(namely xxd and ssh) were missing. This change should fix this by
aborting the prepare step in case if these utilities are not found.

Change-Id: Ic49c11054345964c9785e9c4c791ea33dba4791c
Partial-Bug: 1525967