Fix PATH for VirtualBox when Windows
The support to deploy on a remote machine is broke the
default behaviour on the Windows OS, so now it always
needs to set the $PATH manually before using the scripts.

Also, the new functionality didn't work on the remote
Windows machine because the scripts change the $PATH on
the local system only.

This commit adds some more checks and fixes the local
$PATH setup. It didn't check remote system requirements
so deploy on a remote machine still have some problems.
Added several messages in the scripts.

Change-Id: If5dadc1ece3931f42eac423aae4d7016d45c2040
Closes-Bug: #1534129
Partial-Bug: #1491925