Add listing target
This new 'listing' target puts lists of
files in $(BUILD_DIR) and in local_mirror
to files, so to use them to identify the content of the ISO.

It also finds all fuel related rpm packages that are put into the ISO
and puts their changelog into listing-package-changelog.txt
file. This file can be used to identify SHA sums
of repos that are used to build corresponding packages.

Change-Id: I4738cea6fbd9539b945607943f3bcaf1c0554057
Related-Bug: #1515517