Fix USB installation
The follwing errors during installation from USB stick/image
caused by the attempt to mount partition that already used

  + mount /dev/disk/by-label/Mirantis_Fuel /tmp/source
  mount: /dev/sda1 is write-protected, mounting read-only
  mount: /dev/sda1 is already mounted or /tmp/source busy

Anaconda mounts paritions from 'repo' automatically (not sure
about nfs partitions but there is a code to mount is explicitely),
so it's not possible to re-mount it. However, we can mount it
with 'bind' option to make avaible in another path.

This commit collects all the logic to mount SOURCE path
in one place and adds additional checks to avoid mointing when
SOURCE path set up already.

Change-Id: Idd7f56f23a8bab604d78f2b83f0290e8006482dc
Closes-bug: #1538335