Read ip= from kernel parameters to pass it to fuelmenu on a first boot
We have a default ip= line in kernel parameters setting up the "eth0"
interface. System tests change this line in order to use another
network interface as the admin network interface. It used to work with
7.0 (admin_interface parameter) but now it doesn't.

This patch extracts the device name from the kernel parameters string to
pass it later to fuelmenu on a first boot.

Change-Id: If7f50a7e0e4f9cb3f02bd4e5a5f6922d7c9e8b57
Closes-Bug: #1543080