Use CentOS 7 release for master node
A commit similar to this one was applied (and then reverted because of
some issues) some time ago to switch ISO to CentOS-7.2. When all issues
were resolved we finally have switched to CentOS-7 using intermediate
snapshots that updating constantly. A path to snapshot to use is passed
from CI rather than should be generated inside fuel-main as it was
before.

However, there are community builds that doesn't use snapshotting
mechanism but also should be switched to CentOS-7.

This commit does this, switching ISO to consume the latest CentOS
release [1, 2]

[1]. http://mirror.centos.org/centos/7/
[2]. https://specs.openstack.org/openstack/fuel-specs/specs/9.0/switch-to-centos-7-2.html

Blueprint: switch-to-centos-7-2

Change-Id: I117343c29921238aaef73d7deb3494ea1421e404